requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.6.1
RPi.GPIO==0.6.3
RTIMULib==7.2.1
SPI-Py==1.0
Werkzeug==0.9.6
chardet==2.3.0
codebug-i2c-tether==0.2.3
codebug-tether==0.7.0
colorama==0.3.2
gpiozero==1.3.2
html5lib==0.999
itsdangerous==0.24
mcpi==0.1.1
numpy==1.8.2
pgzero==1.1
pi-rc522==2.2.1
picamera==1.13
picraft==1.0
pifacecommon==4.2.1
pifacedigitalio==3.1.0
pigpio==1.35
pyOpenSSL==0.13.1
pygame==1.9.2a0
pygobject==3.14.0
pyinotify==0.9.4
pyserial==2.6
python-apt==0.9.3.12
python-debian==0.1.27
requests==2.4.3
sense-emu==1.0
sense-hat==2.2.0
six==1.8.0
smbus==1.1
spidev==3.0
twython==3.1.2
urllib3==1.9.1
wheel==0.24.0